requirements.txt
Pillow==9.4.0
requests==2.28.2
streamlit==1.17.0
streamlit_lottie==0.0.3
